+++ scene2d/src/main/kotlin/ktx/scene2d/tree.kt
@@ -11,7 +11,8 @@ import com.badlogic.gdx.scenes.scene2d.ui.Tree
  * @param init will be invoked on the widget. Inlined.
  * @return a new [Tree] instance.
  */
-inline fun tree(style: String = defaultStyle, skin: Skin = Scene2DSkin.defaultSkin,
+inline fun tree(style: String = defaultStyle,
+                skin: Skin = Scene2DSkin.defaultSkin,
                 init: KTreeWidget.() -> Unit) = actor(KTreeWidget(skin, style), init)
 
 /**